+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Refactorings
 
 * ContourMixin: to_polygon (for both 2D and 3D)
+* BSplineCurve2D.point_distance
 * new dataclass EdgeStyle: to be used in several plot methods. simplifying its structure.
 
 ## v0.9.0 [Testing]